feat(stops): integrate GeoJSON feature collection into StopsList context
- Added a new feature collection to the StopsList context, allowing for the transformation of stop data into GeoJSON format.
- Updated StopsListMap to utilize the new features property from context, enhancing the mapping of stops with geographical coordinates.
- Removed redundant GeoJSON transformation logic from StopsListMap for cleaner code structure.
